// src/api/models/RemboursementResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RemboursementResponse = {
    idRemboursement?: string;
    dateFacturation?: string;
    dateComptable?: string;
    referencePaiement?: string;
    banqueDestination?: string;
    dateEcheance?: string;
    montant?: number;
    devise?: string;
    tauxChange?: number;
    motif?: string;
    numeroPiece?: string;
    statut?: string;
    idFacture?: string;
    idClient?: string;
    createdAt?: string;
    updatedAt?: string;
};